test --config help; Keep help tests in one test module
